# Use float for color space conversion