--- a/osu.Framework.Benchmarks/BenchmarkColourInfo.cs
+++ b/osu.Framework.Benchmarks/BenchmarkColourInfo.cs
@@ -1,6 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
 using BenchmarkDotNet.Attributes;
 using osu.Framework.Graphics.Colour;
 using osuTK.Graphics;
@@ -10,24 +11,29 @@ namespace osu.Framework.Benchmarks
     [MemoryDiagnoser]
     public class BenchmarkColourInfo
     {
-        private ColourInfo colourInfo;
+        [ParamsSource(nameof(ColourParams))]
+        public ColourInfo Colour { get; set; }
 
-        [GlobalSetup]
-        public void GlobalSetup()
+        public IEnumerable<ColourInfo> ColourParams
         {
-            colourInfo = ColourInfo.SingleColour(Color4.Transparent);
+            get
+            {
+                yield return ColourInfo.SingleColour(Color4.Transparent);
+                yield return ColourInfo.SingleColour(Color4.Cyan);
+                yield return ColourInfo.SingleColour(Color4.DarkGray);
+            }
         }
 
         [Benchmark]
-        public SRGBColour ConvertToSRGBColour() => colourInfo;
+        public SRGBColour ConvertToSRGBColour() => Colour;
 
         [Benchmark]
-        public Color4 ConvertToColor4() => ((SRGBColour)colourInfo).Linear;
+        public Color4 ConvertToColor4() => ((SRGBColour)Colour).Linear;
 
         [Benchmark]
         public Color4 ExtractAndConvertToColor4()
         {
-            colourInfo.TryExtractSingleColour(out SRGBColour colour);
+            Colour.TryExtractSingleColour(out SRGBColour colour);
             return colour.Linear;
         }
     }