@bdach 2023-05-20
Benchmark appears to confirm gains to the tune of 2x faster in the best cases.

Doing math entirely in single-precision is bound to not produce the exact same results as in double-precision followed by truncation, but I'm not super sure I care too much about that. Any difference that this could cause would be mostly a precision artifact anyway.