@huoyaoyuan 2023-05-21
> Doing math entirely in single-precision is bound to not produce the exact same results as in double-precision followed by truncation

I even considered `Half` as an option since color do not need such precision, but there's no direct math support for math now.

Anyway, the precisions of floating point already differ on different platforms, especially `Math` which delegates to platform CRT.