@bdach 2023-05-21
Pushed additional changes:

- Applied same optimisation to `Colour4` too
- Added a visual test scene with coloured boxes to use to test and compare visual output (doesn't seem to change anything on linux)

Seems ok to me. @peppy double check would be appreciated